Debug functionality is in
In each scene there is an object called _General, at the top of the
hierarch, it has a checkbox for control by kinect, if you turn that off,
you can control things from the keyboard using wasd, shift and control
for movement, and q and e for left and right hand fireballs. If you are
working on a script and need to check whether it is in debug mode, it is
a static variable, just type if(General.kinectControl), which is a bool
and use that.